pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Approximator <alex@nls.la>", "Olha Aliakina <olusiabn@gmail.com"]
include = ["doc/*.md"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
django-rest-knox = "^4.2.0"
djoser = "^2.2.0"
django-cors-headers = "^4.1.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.26.2"
social-auth-core = "^4.4.2"
social-auth-app-django = "^5.2.0"
whitenoise = "^6.5.0"
pillow = "^10.0.0"
daphne = "^4.0.0"
channels = "^4.0.0"
channels-redis = "^4.1.0"
drf-standardized-errors = {extras = ["openapi"], version = "^0.12.5"}
PyYAML = "^6.0.1"
aiohttp = {extras = ["speedups"], version = "^3.9.1"}

[tool.poetry.dev-dependencies]
black = "^23.3.0"
isort = "^5.12.0"
autoflake = "^2.1.1"
pytest-django = "^4.6.0"
pytest-asyncio = "^0.23.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] [%(filename)s:%(lineno)s] %(message)s"
addopts = "--capture=no --show-capture=all"
testpaths = [
    "tests",
]
DJANGO_SETTINGS_MODULE = "config.settings_test"

[tool.black]
line-length = 120
